Extract test env classes
