+++ server/src/test/java/io/spine/server/rejection/given/DelegatingRejectionDispatcherTestEnv.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright 2018, TeamDev. All rights reserved.
+ *
+ * Redistribution and use in source and/or binary forms, with or without
+ * modification, must retain the above copyright notice and the following
+ * disclaimer.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package io.spine.server.rejection.given;
+
+import com.google.common.collect.ImmutableSet;
+import io.spine.core.RejectionClass;
+import io.spine.core.RejectionEnvelope;
+import io.spine.server.entity.rejection.StandardRejections;
+import io.spine.server.rejection.RejectionDispatcherDelegate;
+
+import java.util.Set;
+
+/**
+ * @author Alexander Yevsyukov
+ * @author Dmytro Kuzmin
+ */
+public class DelegatingRejectionDispatcherTestEnv {
+
+    /** Prevents instantiation of this utility class. */
+    private DelegatingRejectionDispatcherTestEnv() {
+    }
+
+    public static final class EmptyRejectionDispatcherDelegate
+            implements RejectionDispatcherDelegate<String> {
+
+        private boolean onErrorCalled;
+        private boolean dispatchCalled;
+
+        @Override
+        public Set<RejectionClass> getRejectionClasses() {
+            return ImmutableSet.of(
+                    RejectionClass.of(StandardRejections.EntityAlreadyArchived.class));
+        }
+
+        @Override
+        public Set<RejectionClass> getExternalRejectionClasses() {
+            return ImmutableSet.of();
+        }
+
+        @Override
+        public Set<String> dispatchRejection(RejectionEnvelope envelope) {
+            dispatchCalled = true;
+            return ImmutableSet.of(toString());
+        }
+
+        @Override
+        public void onError(RejectionEnvelope envelope, RuntimeException exception) {
+            onErrorCalled = true;
+        }
+
+        public boolean onErrorCalled() {
+            return onErrorCalled;
+        }
+
+        public boolean dispatchCalled() {
+            return dispatchCalled;
+        }
+    }
+}
